Update index.d.ts with errorCode argument
The errorCode option creates an error in TypeScript because it is not part of the index.d.ts file. However ignoring the build errors show that the errorCode is applied to the error result as excepted. Adding errorCode as argument to the methods in index.d.ts, will remove the build error in TypeScript when using the errorCode property.